requirements.txt
bitcoinlib
Pillow
openai
python-dotenv








